Fix compiler warnings in apps and plugins
- Replace deprecated Qt API (type() -> typeId(), QVariant::String -> QMetaType::QString)
- Change WebEngineBackend constructor to pass-by-value with std::move
- Change dynamic_cast usages where appropriate for downcasting
- Use auto* for pointer results from dynamic_cast
- Add explicit type conversions for narrowing conversions (qint64 to double)
- Replace bool conversion warnings with explicit nullptr comparisons
- Add [[nodiscard]] attribute to RemoteService::isAlive()
- Update WebEngineBackend.h to match implementation signature

+++ apps/EKiosk/src/SchedulerTasks/TimeSync.cpp
@@ -195,7 +195,7 @@ void TimeSync::timeOffsetReceived(qint64 aLocalTimeOffset) {
         if (qAbs(aLocalTimeOffset / 1000) > CTimeSync::DesyncLimit) {
             toLog(LogLevel::Normal,
                   QString("Required time synchronization. Offset %1 sec.")
-                      .arg(aLocalTimeOffset / 1000., 0, 'f', 3));
+                      .arg(double(aLocalTimeOffset) / 1000., 0, 'f', 3));
 
             try {
                 ISysUtils::setSystem_Time(QDateTime::currentDateTime().addMSecs(aLocalTimeOffset));
@@ -211,7 +211,7 @@ void TimeSync::timeOffsetReceived(qint64 aLocalTimeOffset) {
         } else {
             toLog(LogLevel::Normal,
                   QString("Time synchronization is not required.  Offset %1 sec.")
-                      .arg(aLocalTimeOffset / 1000., 0, 'f', 3));
+                      .arg(double(aLocalTimeOffset) / 1000., 0, 'f', 3));
         }
     }
 }
